refactor: 修复下载文件的部分问题，将部分功能独立到Lib，以方便插件调用
