+++ Lib/__init__.py
@@ -14,6 +14,7 @@ import Lib.BotController as BotController
 import Lib.Configs as Configs
 import Lib.FileCacher as FileCacher
 import Lib.ThreadPool as ThreadPool
+import Lib.ListeningServer as ListeningServer
 
 
 VERSION = "2.0.0-dev"
